Fix for issues found with VC 17 link time code generation
Added scratch project to isolate problem investigations. Fixed compile test not to trigger not reached cases because something above it resolved to an exception at compile time. Also factored out the test for the modulus friend function because it was
being a problem, and causing the linker to hang. Oddly enough, the change at line 6750 fixes it.