Fix implicit copy constructor clang 10.0 diagnostic (-Weverything, that we use).
.../safeint.hpp.original:5741:35: error: definition of implicit copy constructor for 'SafeInt<long, SafeIntInternal::SafeIntExceptionHandler<SafeIntException> >' is deprecated because it has a user-declared copy assignment operator [-Werror,-Wdeprecated-copy]
    _CONSTEXPR14 SafeInt< T, E >& operator =( const SafeInt< T, E >& rhs ) SAFEINT_NOTHROW
                                  ^
.../safeint.hpp.original:6536:12: note: in implicit copy constructor for 'SafeInt<long, SafeIntInternal::SafeIntExceptionHandler<SafeIntException> >' first required here
    return SafeInt<ptrdiff_t, SafeIntDefaultExceptionHandler>( p1 - p2 );
           ^

There are a few things to consider here.

We could do this:

ifdef __clang__
pragma clang diagnostic ignored "-Wdouble-promotion"
pragma clang diagnostic ignored "-Wdeprecated-copy"
endif

...

ifdef __clang__
pragma clang diagnostic pop
endif

is very safe and tempting.

We can remove the operator= and let the compiler supply the default. This seems like a good idea.

We can provide copy constructor, that the compiler was providing anyway:
   _CONSTEXPR11 SafeInt(const SafeInt& other) SAFEINT_NOTHROW : m_int(other.m_int) { }

Experimentation suggests the compiler-supplied copy constructor is not explicit.
i.e. this program:

```
class A
{
public:

A(int) { }

explicit
A(const A&) {}

void operator=(const A&) { } // warning here, depending on if previous present

};

A f1();
A f1()
{
 return A(1);
}
```

add/remove the copy constructor, and add/remove explicit.
It does compile, with warning, w/o the copy constructor.
It does not compiler if the copy constructor is explicit.

I do need to remove the wisdom about when to provide:
 default constructor
 or copy constructor
 or operator=

When to let compiler provide, and when to say =default.

Here I have provided =default for C++>=11 and manual for C++98.

I wonder if providing none is best.